Course_Projects/06_capstone_project/backend/requirements.txt
aniso8601==9.0.1
authlib==1.3.0
astroid==3.0.2
Click==8.1.3
ecdsa==0.18.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
future==0.17.1
gunicorn==21.2.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jq==1.6.0
lazy-object-proxy==1.10.0
mccabe==0.7.0
MarkupSafe==2.1.1
pylint==3.0.3
python-jose==3.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pytz==2022.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
typed-ast==1.4.2
Werkzeug==3.0.1
wrapt==1.14.1

